Add files via upload
A To-Do List is a simple application that helps you keep track of tasks you want to complete. You can add new tasks, view all the tasks, and remove the ones you've finished. This program saves your tasks in a file so that even if you close it, your tasks will still be there the next time you open it.